+++ src/features/posture/timer/components/Display.tsx
